requirements.txt
customtkinter==5.2.0
pefile==2022.5.30
Pillow==9.5.0
